requirements.txt
# Core dependencies with pinned versions
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.26.0
boto3==1.34.0
pyyaml==6.0.1
prometheus-client==0.19.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
structlog==24.1.0
aiofiles==23.2.1
cryptography==42.0.0

# Multi-PAR supervisor dependencies
psutil==5.9.8
grpcio==1.60.0
grpcio-tools==1.60.0